Fix Commit

+++ contracts/plugins/assets/bancor/BancorV3FiatCollateral.sol
@@ -121,7 +121,7 @@ contract BancorV3FiatCollateral is Collateral {
     /// Claim rewards earned by holding a balance of the ERC20 token
     /// @dev delegatecall
     function claimRewards() external virtual override {
-        uint192 programId = StandardRewards.latestProgramId(0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48);
-        StandardRewards.claimRewards(programId);
+        uint192 programId = uint192(StandardRewards.latestProgramId(0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48));
+        StandardRewards.claimRewards(uint256(programId));
     }
 }
